+++ src/me/hkgumbs/tictactoe/main/java/configuration/PlayersConfiguration.java
@@ -1,9 +1,10 @@
-package me.hkgumbs.tictactoe.main.java.simulation;
+package me.hkgumbs.tictactoe.main.java.configuration;
 
 import me.hkgumbs.tictactoe.main.java.board.Board;
 import me.hkgumbs.tictactoe.main.java.player.Human;
 import me.hkgumbs.tictactoe.main.java.player.Minimax;
 import me.hkgumbs.tictactoe.main.java.player.Player;
+import me.hkgumbs.tictactoe.main.java.simulation.Simulation;
 
 import java.io.InputStream;
 import java.io.OutputStream;
