+++ src/me/hkgumbs/tictactoe/main/java/utility/Extractor.java
@@ -1,4 +1,4 @@
-package me.hkgumbs.tictactoe.main.java.simulation;
+package me.hkgumbs.tictactoe.main.java.utility;
 
 import java.util.List;
