+++ src/me/hkgumbs/tictactoe/test/java/board/SquareBoardTest.java
@@ -11,7 +11,7 @@ import java.util.Set;
 
 import static org.junit.Assert.*;
 
-public class BoardTest {
+public class SquareBoardTest {
 
     Board board;
